Actually allow sending partition key from producer
I was sending the partition key around explicitly in the inner piping,
but the it was not possible specify the partition key when sending a
record from the exposed consumer interface.

Now you can manually send the partition key from the
KinesisProducer. If you don't specify one, it will fall back to using
a provided partitioner, which defaults to creating random partition
keys.